+++ src/snowcli/cli/snowpark/common.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Optional
 
 from snowcli.cli.common.sql_execution import SqlExecutionMixin
 from snowcli.cli.constants import ObjectType
-from snowcli.utils import generate_deploy_stage_name
+from snowcli.utils.utils import generate_deploy_stage_name
 from snowflake.connector.cursor import SnowflakeCursor
 
 DEFAULT_RUNTIME = "3.8"
