+++ tests/test_zipper.py
@@ -1,7 +1,7 @@
 from zipfile import ZipFile
 
 from snowcli.cli.snowpark.commands import _replace_handler_in_zip
-from snowcli.zipper import add_file_to_existing_zip, zip_dir
+from snowcli.utils.zipper import add_file_to_existing_zip, zip_dir
 
 from tests.testing_utils.fixtures import *
